Block statistics (#338)
* refactor search dialog

move range selection into own widget to be able to use it also in other dialog

* refactor search file names

to get better separation:
Widgets are part of a dialog
Dialogs are shown
Plugins are linked to a dialog via an plugin interface

* add statistic for Block

also much refactoring to reuse code from search

* fix linker errors & warnings

* write results to CSV file

* allow regular expressions as search name

(additional to exact match as before)

* add tool-tip for result image

+++ identifier/definitionmanager.cpp
@@ -38,7 +38,7 @@ static quint32 stableHash(const quint8 *data, int size)
 static quint32 stableHash(const char *data, int size = -1)
 {
   if (size < 0)
-    size = strlen(data);
+    size = static_cast<int>(strlen(data));
   return stableHash(reinterpret_cast<const quint8 *>(data), size);
 }
